Change AppLayout to allow any kind element in the menu slot
Moved logic related to the default menu to AppLayoutMenu
