+++ vaadin-app-layout-flow/src/main/java/com/vaadin/flow/component/applayout/AbstractAppRouterLayout.java
@@ -27,14 +27,16 @@ import com.vaadin.flow.router.RouterLayout;
  * Basic usage involves extending this class and implementing
  * the {@code configure} method.
  *
- * @see AbstractAppRouterLayout#configure(AppLayout)
+ * @see AbstractAppRouterLayout#configure(AppLayout, AppLayoutMenu)
  */
 public abstract class AbstractAppRouterLayout implements RouterLayout {
 
     private AppLayout appLayout = new AppLayout();
 
+    private AppLayoutMenu appLayoutMenu = appLayout.createMenu();
+
     protected AbstractAppRouterLayout() {
-        configure(getAppLayout());
+        configure(getAppLayout(), getAppLayoutMenu());
     }
 
     /**
@@ -43,13 +45,14 @@ public abstract class AbstractAppRouterLayout implements RouterLayout {
      *
      * @param appLayout {@link AppLayout} to configure.
      */
-    protected abstract void configure(AppLayout appLayout);
+    protected abstract void configure(AppLayout appLayout,
+        AppLayoutMenu appLayoutMenu);
 
     /**
      * This hook is called when a navigation is being made into a route
      * which has this router layout as its parent layout.
      *
-     * @param route route that is being navigated to
+     * @param route   route that is being navigated to
      * @param content the content component
      */
     protected void onNavigate(String route, HasElement content) {
@@ -57,13 +60,13 @@ public abstract class AbstractAppRouterLayout implements RouterLayout {
 
     @Override
     public void showRouterLayoutContent(HasElement content) {
-        final String target = UI.getCurrent().getRouter().getUrl(
-                content.getElement().getComponent().get().getClass());
+        final String target = UI.getCurrent().getRouter()
+            .getUrl(content.getElement().getComponent().get().getClass());
 
         onNavigate(target, content);
 
-        getAppLayout().getMenuItemTargetingRoute(target)
-                .ifPresent(getAppLayout()::selectMenuItem);
+        getAppLayoutMenu().getMenuItemTargetingRoute(target)
+            .ifPresent(getAppLayoutMenu()::selectMenuItem);
         getAppLayout().setContent(content.getElement());
     }
 
@@ -77,5 +80,9 @@ public abstract class AbstractAppRouterLayout implements RouterLayout {
      */
     public AppLayout getAppLayout() {
         return appLayout;
+    }
+
+    public AppLayoutMenu getAppLayoutMenu() {
+        return appLayoutMenu;
     }
 }
