+++ vaadin-app-layout-flow/src/main/java/com/vaadin/flow/component/applayout/AppLayoutMenu.java
@@ -18,6 +18,8 @@ package com.vaadin.flow.component.applayout;
  */
 
 import com.vaadin.flow.component.AttachNotifier;
+import com.vaadin.flow.component.Component;
+import com.vaadin.flow.component.ComponentEventListener;
 import com.vaadin.flow.component.HasElement;
 import com.vaadin.flow.component.tabs.Tabs;
 import com.vaadin.flow.dom.Element;
@@ -26,19 +28,18 @@ import java.util.Objects;
 import java.util.Optional;
 
 /**
- * Tabs for AppLayout.
+ * Menu to be used with AppLayout. Provides clicable tabs that can be used for routing or individual actions.
  */
-class AppLayoutMenu implements HasElement, AttachNotifier {
+public class AppLayoutMenu implements HasElement, AttachNotifier {
 
     private final Tabs tabs = new Tabs();
     private AppLayoutMenuItem selectedMenuItem;
 
     /**
-     * Initializes a new app layout with a default menu.
+     * Default constructor.
      */
-    AppLayoutMenu() {
-        getElement().setAttribute("slot", "menu");
-        getElement().setAttribute("theme", "minimal");
+    public AppLayoutMenu() {
+        tabs.getElement().setAttribute("theme", "minimal");
 
         tabs.addAttachListener(attachEvent -> {
             tabs.setSelectedTab(selectedMenuItem);
@@ -69,7 +70,7 @@ class AppLayoutMenu implements HasElement, AttachNotifier {
      *
      * @param menuItems items to set
      */
-    void setMenuItems(AppLayoutMenuItem... menuItems) {
+    public void setMenuItems(AppLayoutMenuItem... menuItems) {
         clearMenuItems();
         addMenuItems(menuItems);
     }
@@ -79,14 +80,93 @@ class AppLayoutMenu implements HasElement, AttachNotifier {
      *
      * @param menuItems items to add
      */
-    void addMenuItems(AppLayoutMenuItem... menuItems) {
+    public void addMenuItems(AppLayoutMenuItem... menuItems) {
         tabs.add(menuItems);
+    }
+
+    /**
+     * Constructs a new object with the given title.
+     *
+     * @param title the title to display
+     */
+    public AppLayoutMenuItem addMenuItem(String title) {
+        return addAndReturn(new AppLayoutMenuItem(title));
+    }
+
+    /**
+     * Constructs a new object with the given icon.
+     *
+     * @param icon the icon to display
+     */
+    public AppLayoutMenuItem addMenuItem(Component icon) {
+        return addAndReturn(new AppLayoutMenuItem(icon));
+    }
+
+    /**
+     * Constructs a new object with the given icon and title.
+     *
+     * @param icon  the icon to display
+     * @param title the title to display
+     */
+    public AppLayoutMenuItem addMenuItem(Component icon, String title) {
+        return addAndReturn(new AppLayoutMenuItem(icon, title));
+    }
+
+    /**
+     * Constructs a new object with the given icon, title and route.
+     *
+     * @param icon  the icon to display
+     * @param title the title to display
+     * @param route the route to navigate on click
+     */
+    public AppLayoutMenuItem addMenuItem(Component icon, String title,
+        String route) {
+        return addAndReturn(new AppLayoutMenuItem(icon, title, route));
+    }
+
+    /**
+     * Constructs a new object with the given icon and click listener.
+     *
+     * @param icon     the icon to display
+     * @param listener the menu item click listener
+     */
+    public AppLayoutMenuItem addMenuItem(Component icon,
+        ComponentEventListener<MenuItemClickEvent> listener) {
+        return addAndReturn(new AppLayoutMenuItem(icon, listener));
+    }
+
+    /**
+     * Constructs a new object with the given title and click listener.
+     *
+     * @param title    the title to display
+     * @param listener the menu item click listener
+     */
+    public AppLayoutMenuItem addMenuItem(String title,
+        ComponentEventListener<MenuItemClickEvent> listener) {
+        return addAndReturn(new AppLayoutMenuItem(title, listener));
+    }
+
+    /**
+     * Constructs a new object with the given icon, title and click listener.
+     *
+     * @param icon     the icon to display
+     * @param title    the title to display
+     * @param listener the menu item click listener
+     */
+    public AppLayoutMenuItem addMenuItem(Component icon, String title,
+        ComponentEventListener<MenuItemClickEvent> listener) {
+        return addAndReturn(new AppLayoutMenuItem(icon, title, listener));
+    }
+
+    private AppLayoutMenuItem addAndReturn(AppLayoutMenuItem item) {
+        addMenuItems(item);
+        return item;
     }
 
     /**
      * Removes menu item from the menu
      */
-    void removeMenuItem(AppLayoutMenuItem menuItem) {
+    public void removeMenuItem(AppLayoutMenuItem menuItem) {
         if (Objects.equals(this.selectedMenuItem, menuItem)) {
             this.selectedMenuItem = null;
         }
@@ -104,21 +184,30 @@ class AppLayoutMenu implements HasElement, AttachNotifier {
     /**
      * Selects a menu item.
      */
-    void selectMenuItem(AppLayoutMenuItem menuItem) {
+    public void selectMenuItem(AppLayoutMenuItem menuItem) {
         selectedMenuItem = menuItem;
         tabs.setSelectedTab(menuItem);
     }
 
-    Optional<AppLayoutMenuItem> getMenuItemTargetingRoute(String route) {
+    /**
+     * Gets the first {@link AppLayoutMenuItem} targeting a route.
+     */
+    public Optional<AppLayoutMenuItem> getMenuItemTargetingRoute(String route) {
         Objects.requireNonNull(route, "Route can not be null");
         return tabs.getChildren().map(AppLayoutMenuItem.class::cast)
             .filter(e -> route.equals(e.getRoute())).findFirst();
     }
 
-    AppLayoutMenuItem getSelectedMenuItem() {
+    /**
+     * Gets the currently selected menu item.
+     */
+    public AppLayoutMenuItem getSelectedMenuItem() {
         return selectedMenuItem;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public Element getElement() {
         return tabs.getElement();
