+++ vaadin-app-layout-integration-tests/src/main/java/com/vaadin/flow/component/applayout/examples/AppRouterLayout.java
@@ -3,6 +3,7 @@ package com.vaadin.flow.component.applayout.examples;
 import com.vaadin.flow.component.UI;
 import com.vaadin.flow.component.applayout.AbstractAppRouterLayout;
 import com.vaadin.flow.component.applayout.AppLayout;
+import com.vaadin.flow.component.applayout.AppLayoutMenu;
 import com.vaadin.flow.component.applayout.AppLayoutMenuItem;
 import com.vaadin.flow.component.html.Span;
 import com.vaadin.flow.component.icon.VaadinIcon;
@@ -21,19 +22,19 @@ import java.util.stream.IntStream;
 public class AppRouterLayout extends AbstractAppRouterLayout {
 
     @Override
-    protected void configure(AppLayout appLayout) {
+    protected void configure(AppLayout appLayout, AppLayoutMenu appLayoutMenu) {
         appLayout.setBranding(new Span("Vaadin").getElement());
 
-        appLayout.addMenuItems(generateMenuItems(
+        appLayoutMenu.addMenuItems(generateMenuItems(
             i -> new AppLayoutMenuItem(VaadinIcon.SAFE_LOCK.create(),
                 "Action " + i, e -> Notification
                 .show(e.getSource().getTitle() + " executed!"))));
 
-        appLayout.addMenuItems(generateMenuItems(
+        appLayoutMenu.addMenuItems(generateMenuItems(
             i -> (new AppLayoutMenuItem(VaadinIcon.LOCATION_ARROW.create(),
                 "Page " + i, "Page" + i))));
 
-        appLayout.addMenuItems(
+        appLayoutMenu.addMenuItems(
             new AppLayoutMenuItem(VaadinIcon.HOME.create(), "Home", ""),
             new AppLayoutMenuItem(VaadinIcon.USER.create(), "Logout",
                 e -> UI.getCurrent().navigate("LoggedOut")));
